backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
bcrypt==4.2.0
pymysql==1.1.1
PyJWT==2.9.0
cryptography==43.0.1

pytest~=8.3.3
starlette~=0.41.2
passlib~=1.7.4
alembic~=1.13.3
pydantic-settings~=2.6.0
psutil~=6.1.0
requests~=2.32.3